fix: converted dialog.js to dialog.jsx
